Core update
Added Text class and error screen state when it fails loading map file.

I've updated resManager so it is able now to keep variables, which is useful for sending messages over states and also for having global constants.
Some minor changes in the code, removed unused imports and so on.